requirements.txt
mysql.connector
pandas
numpy
pymysql
google-generativeai
streamlit
python-dotenv